requirements.txt
aiocoap
pynacl
cbor2
